Make InMemoryRepository class To do divide DB dependency
