+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -10,7 +10,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/tasks")
 public class TaskController {
-    private TaskService taskService;
+    private final TaskService taskService;
 
     public TaskController(TaskService taskService) {
         this.taskService = taskService;
